requirements.txt
torch>=1.2
torchvision
visdom
numpy
pillow
scikit-learn
